this slices our first build step out of the monolithic grunt build.

improvements:

 - faster less build
 - no useless css files for modules in debug folder, just one
   `index.css`
 - multiple targets possible (so we don't need to build from
   `dist/debug` in the future

performance for less build & concat:

```
old:
 - 762ms -- 757ms + 5ms (separate concat)

new:
 - 597ms
```

why is it faster?

1. previous buildstep wrote css files on the disk to read & concat
it later

2. grunt.readfile / grunt.readjson are slow compared to the native
require function which also supports caching

needs node 4.x like our travis builds