@benkeen 2015-12-01
This looks great, Robert. Works fine. Nice to see some ES6 code in Fauxton. :+1: +1